requirements.txt
mutagen>=1.47
watchdog>=4.0
python-dotenv>=1.0
PySide6>=6.7
tqdm>=4.66
rich>=13.7
pydantic>=2.7
typer>=0.12
PyYAML>=6.0
ffmpeg-python>=0.2
pyacoustid>=1.2.2
musicbrainzngs>=0.7.1
pytest>=8.2
httpx>=0.27
